Refactored shortcut modal, began settings implementation, added base README
