+++ src/electron/main.ts
@@ -1,14 +1,27 @@
 import { app, globalShortcut, BrowserWindow, ipcMain } from "electron";
 import path from "path";
-import fs from "fs";
+import Store from "electron-store";
 
 import * as utils from "./utils.js";
 import * as constants from "./constants.js";
 import { ShortcutProps } from "./types.js";
+import { ShortcutManager } from "./shortcut-manager.js";
+
+type Settings = {
+  notificationBannersEnabled: boolean;
+  notificationSoundsEnabled: boolean;
+};
+
+const store = new Store<Settings>({
+  defaults: {
+    notificationBannersEnabled: true,
+    notificationSoundsEnabled: true,
+  },
+});
 
 let mainWindow: BrowserWindow | null;
 
-let shortcutList: Record<string, ShortcutProps> = {};
+const shortcutManager = new ShortcutManager(store);
 
 const createWindow = () => {
   // Create the browser window.
@@ -44,26 +57,6 @@ const createWindow = () => {
 app.whenReady().then(() => {
   createWindow();
 
-  // Get list of shortvuts from file (if exists)
-  if (fs.existsSync(constants.SHORTCUT_LIST_PATH)) {
-    try {
-      const rawData = fs.readFileSync(constants.SHORTCUT_LIST_PATH, "utf-8");
-      shortcutList = JSON.parse(rawData);
-      // Re-register all shortcuts from the list on startup.
-      for (const id in shortcutList) {
-        const shortcut = shortcutList[id];
-        if (shortcut) {
-          utils.registerShortcut(shortcut);
-        }
-      }
-    } catch (error) {
-      console.log("unable to read file: ", error);
-    }
-  } else {
-    console.log("file doesn't exist");
-    fs.writeFileSync(constants.SHORTCUT_LIST_PATH, "{}");
-  }
-
   globalShortcut.register(constants.startCommand, () => {
     if (!mainWindow) {
       createWindow();
@@ -78,61 +71,31 @@ app.whenReady().then(() => {
 
   // Define IPC event handlers
   ipcMain.handle("add-shortcut", (event, data: ShortcutProps) => {
-    console.log("adding shortcut: ", data);
-    if (globalShortcut.isRegistered(data.accelerator)) {
-      const message = `Shortcut ${data.accelerator} is already registered.`;
-      console.log(message);
-      return { success: false, message };
-    }
-
-    const success = utils.registerShortcut(
-      data
-    );
-
-    if (success) {
-      shortcutList[data.id] = data;
-      utils.saveShortcutList(shortcutList);
-      return { success: true, message: "Shortcut registered successfully" };
-    } else {
-      const message = `Failed to register shortcut: ${data.accelerator}. It might be an invalid combination.`;
-      console.log(message);
-      return { success: false, message };
-    }
+    return shortcutManager.addShortcut(data);
   });
 
   ipcMain.handle("get-shortcuts", () => {
-    return shortcutList;
+    return shortcutManager.getShortcuts();
   });
 
   ipcMain.handle("delete-shortcut", (event, shortcutId) => {
-    if (!shortcutId || !shortcutList[shortcutId]) {
-      console.error("Invalid shortcut ID for deletion:", shortcutId);
-      return { success: false, message: "Invalid shortcut ID." };
-    }
-
-    const shortcutToDelete = shortcutList[shortcutId];
-    
-    // Unregister from Electron
-    globalShortcut.unregister(shortcutToDelete.accelerator);
-
-    // Remove from in-memory list
-    delete shortcutList[shortcutId];
-
-    // Persist the updated list
-    try {
-      fs.writeFileSync(constants.SHORTCUT_LIST_PATH, JSON.stringify(shortcutList, null, 2));
-      console.log(`Shortcut ${shortcutToDelete.name} deleted and list saved.`);
-    } catch (error) {
-      console.error("Failed to save shortcut list after deletion:", error);
-      // If saving fails, you might want to rollback the in-memory deletion
-      // For simplicity here, we'll just log the error.
-      shortcutList[shortcutId] = shortcutToDelete; // Rollback
-      globalShortcut.register(shortcutToDelete.accelerator, () => {}); // Re-register (action is lost)
-      return { success: false, message: "Failed to update shortcut file." };
-    }
-
-    return { success: true, message: "Shortcut deleted successfully." };
+    return shortcutManager.deleteShortcut(shortcutId);
   });
+
+  ipcMain.handle("settings:get", () => {
+    return {
+      notificationBannersEnabled: store.get("notificationBannersEnabled"),
+      notificationSoundsEnabled: store.get("notificationSoundsEnabled"),
+    };
+  });
+
+  ipcMain.handle(
+    "settings:update",
+    (event, { key, value }: { key: keyof Settings; value: boolean }) => {
+      store.set(key, value);
+      return { success: true };
+    }
+  );
 
   ipcMain.handle("play-shortcut-sound", () => {
     // Send a message to the renderer process to play the sound
