+++ src/electron/word-count-popup.html
@@ -1,0 +1,91 @@
+<!DOCTYPE html>
+<html>
+  <head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Text Analysis</title>
+    <style>
+      body {
+        margin: 0;
+        padding: 24px;
+        font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, sans-serif;
+        background: white;
+        color: #1f2937;
+        border-radius: 12px;
+        box-shadow: 0 10px 25px rgba(0, 0, 0, 0.15);
+        user-select: none;
+        overflow: hidden;
+      }
+      .container {
+        text-align: center;
+      }
+      .title {
+        font-size: 20px;
+        font-weight: 600;
+        margin-bottom: 24px;
+        color: #111827;
+      }
+      .stats-container {
+        display: flex;
+        gap: 16px;
+        justify-content: center;
+        margin-bottom: 20px;
+      }
+      .stat-card {
+        flex: 1;
+        padding: 16px;
+        background: #f9fafb;
+        border: 1px solid #e5e7eb;
+        border-radius: 8px;
+        box-shadow: 0 1px 3px rgba(0, 0, 0, 0.1);
+      }
+      .stat-label {
+        font-size: 14px;
+        color: #6b7280;
+        margin-bottom: 8px;
+        font-weight: 500;
+      }
+      .stat-value {
+        font-size: 28px;
+        font-weight: bold;
+        color: #111827;
+      }
+      .close-btn {
+        width: 100%;
+        padding: 8px 16px;
+        background: #f3f4f6;
+        border: 1px solid #e5e7eb;
+        color: #6b7280;
+        border-radius: 6px;
+        cursor: pointer;
+        font-size: 14px;
+        transition: all 0.2s;
+      }
+      .close-btn:hover {
+        background: #e5e7eb;
+        color: #374151;
+      }
+    </style>
+  </head>
+  <body>
+    <div class="container">
+      <div class="title">📊 Text Analysis</div>
+      <div class="stats-container">
+        <div class="stat-card">
+          <div class="stat-label">Words</div>
+          <div class="stat-value" id="wordCount"></div>
+        </div>
+        <div class="stat-card">
+          <div class="stat-label">Characters</div>
+          <div class="stat-value" id="charCount"></div>
+        </div>
+      </div>
+      <button class="close-btn" onclick="window.close()">Close</button>
+    </div>
+    <script>
+      const urlParams = new URLSearchParams(window.location.search);
+      document.getElementById('wordCount').innerText = urlParams.get('wordCount');
+      document.getElementById('charCount').innerText = urlParams.get('charCount');
+    </script>
+  </body>
+</html>
